Added WebApi project to play around and off course a related test project
Added only one controller, named BankTransfer. I don't want to allow
Post or Delete, because the Transaction are created from an importer. I
will allow Put, so someone can alter the BankTransferExtensions. However
I came to the conclusion that an Id to identify the transactions would
be best. (otherwise you have to compare a lot of other properties to
make a match) I excluded the routing files from Code Coverage and
managed to cover the related part for 100%. I switched from VS2010 to
VS2017 in order to do so. (but badly miss ReSharper)